return available files in GetMountFiles alongside any error
When Readdir returns an error, it may still have a list of files read up until that point.
Previously we returned early if there was an error, rather than returning those files.
This was prevalent when looking in certain directories such as "/etc" of an Ubuntu desktop.
Output should not match results of "ls -la" in mounted directory